Move santise_doc method from conftest to testutils.
